refactor: redis 관련 모듈을 Infra로 이동
